Implemented ContainerCell.
Container cell is a subclass of UITableViewController that is reused by table view. But it doesn't contain any subviews at start. View that we need to add to ContainerCell is stored inside ContainerCellModel. This approach let us prevent reusing mechanism for this views. We simply add needed subview on cell reuse. Can be useful for Tables with static data, difficult layout, but small amount of cells.